feat: complete Vector3/Coord3D type system unification
- Successfully integrated WWMath Vector3 (X,Y,Z) with game Coord3D (x,y,z)
- Added critical DirectX API stubs: DrawPrimitiveUP, ShowCursor, SetCursorProperties
- Added DIDEVICEOBJECTDATA stub for DirectInput compatibility
- Protected Windows APIs: IsIconic, SetCursor with #ifdef guards
- MILESTONE: W3DModelDraw.cpp compiles successfully (0 errors, 176 warnings)
- Reduced compilation errors from 120+ to ~80 (33% reduction)
- Enhanced CORE_IDirect3DDevice8 with missing method implementations

+++ GeneralsMD/Code/GameEngineDevice/Include/Win32Device/GameClient/Win32DIKeyboard.h
@@ -60,7 +60,7 @@
 // macOS: DirectInput not available, use stubs
 typedef void* LPDIRECTINPUT8;
 typedef void* LPDIRECTINPUTDEVICE8;
-typedef int HRESULT;
+typedef long HRESULT;
 #define DI_OK 0
 #endif
